create services

+++ src/main/java/com/example/demo/Reponsitory/RepositoryStudent.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository // Repository annotation help access and interact with database
-public interface RepositoryStudent extends JpaRepository<UniversityStudent, Integer> {
+public interface RepositoryStudent extends JpaRepository<UniversityStudent, Long> {
 }
